Fixed potiential data manager crashes

+++ src/main/java/com/legacy/aether/entities/passive/mountable/EntityAerbunny.java
@@ -116,7 +116,7 @@ public class EntityAerbunny extends EntityAetherAnimal
 
     public int getPuffiness()
     {
-        return this.dataManager.get(PUFF);
+        return this.dataManager.get(PUFF).intValue();
     }
 
     public void setPuffinessClient(int i)
